Add dual target rules to makefile, using C++ on both host and device

+++ py/modbuiltins.c
@@ -146,7 +146,7 @@ static mp_obj_t mp_builtin_chr(mp_obj_t o_in) {
     #else
     mp_int_t ord = mp_obj_get_int(o_in);
     if (0 <= ord && ord <= 0xff) {
-        uint8_t str[1] = {ord};
+        uint8_t str[1] = {(uint8_t)ord};
         return mp_obj_new_str_via_qstr((char *)str, 1);
     } else {
         mp_raise_ValueError(MP_ERROR_TEXT("chr() arg not in range(256)"));
